refactor(core): consolidate MockHttpClient into shared fixtures
Deduplicate the triplicated MockHttpClient (error-map-v2-cleanup, Item 2):

- Unit side: new #[cfg(test)] pub(crate) test_fixtures module in
  webfang_core with a single MockHttpClient; port.rs and domain/ports.rs
  now import it instead of defining local copies (MockHttpClientPort
  renamed at its two usage sites; MockPersistencePort untouched).
- Integration side: new tests/common/mock_http.rs wired via flat
  #[path] (repo convention per cli_binary_test.rs). Deliberately NOT
  routed through tests/common/mod.rs, which is orphaned and references
  the unresolvable webfang:: bin-only crate.

Behavior-identical relocation: port.rs 5/5, domain/ports 4/4, and
scraper_service_test 45/45 stay green unchanged.
